Routes layout

+++ BRASS/Models/RouteStops.cs
@@ -9,8 +9,8 @@ namespace BRASS.Models
         [Key]
         public int StopId { get; set; }
         public int StopNumber { get; set; }
-        public double Longitude { get; set; }
-        public double Lattitude { get; set; }
+        public decimal Longitude { get; set; }
+        public decimal Lattitude { get; set; }
         public int RouteId { get; set; }
     }
 }
